refactor: keep the stream summary out of the generation row and track.ts
Move the summary and log-trimming helpers to src/text/streamSummary.ts;
track.ts only calls them. Drop the errorDetails column write: the summary
lives in the error_event upstream body, and generation_event_v2 is unchanged.
